+++ src/Responses/ResponseInterface.php
@@ -10,17 +10,17 @@ interface ResponseInterface
 {
     /**
      * Converts the response to an array
-     * @param Config $config The configuration instance
+     * @param Config|null $config The configuration instance
      * @return array The array representation of the response
      */
-    public function toArray(Config $config): array;
+    public function toArray(?Config $config = null): array;
 
 
 
     /**
      * Checks if the response is valid
-     * @param Config $config The configuration instance
+     * @param Config|null $config The configuration instance
      * @return bool True if the response is valid, false otherwise
      */
-    public function isValid(Config $config): bool;
+    public function isValid(?Config $config = null): bool;
 }